// apps/auth-admin/frontend/src/types/layout.d.ts
export type MenuMode = 'static' | 'overlay' | 'horizontal' | 'slim' | 'slim-plus' | 'reveal' | 'drawer';

export type ColorScheme = 'light' | 'dark';

export type MenuProfilePosition = 'start' | 'end';

/* Breadcrumb Types */
export interface AppBreadcrumbProps {
    className?: string;
}

export interface Breadcrumb {
    labels?: string[];
    to?: string;
}

export interface BreadcrumbItem {
    label: string;
    to?: string;
    items?: BreadcrumbItem[];
}

/* Context Types */
export type LayoutState = {
    staticMenuDesktopInactive: boolean;
    overlayMenuActive: boolean;
    overlaySubmenuActive: boolean;
    configSidebarVisible: boolean;
    staticMenuMobileActive: boolean;
    menuHoverActive: boolean;
    sidebarActive: boolean;
    anchored: boolean;
    topbarMenuActive: boolean;
    menuProfileActive: boolean;
    rightMenuActive: boolean;
    profileSidebarVisible: boolean;
    resetMenu: boolean;
};

export type LayoutConfig = {
    ripple: boolean;
    inputStyle: string;
    menuMode: MenuMode;
    menuTheme: string;
    colorScheme: ColorScheme;
    scale: number;
    desktopMenuActive: boolean;
    mobileMenuActive: boolean;
    mobileTopbarActive: boolean;
    menuProfilePosition: MenuProfilePosition;
    componentTheme: string;
    topbarTheme: string;
};

export interface LayoutContextProps {
    layoutConfig: LayoutConfig;
    setLayoutConfig: Dispatch<SetStateAction<LayoutConfig>>;
    layoutState: LayoutState;
    setLayoutState: Dispatch<SetStateAction<LayoutState>>;
    onMenuProfileToggle: () => void;
    onMenuToggle: () => void;
    onTopbarMenuToggle: () => void;
    showRightSidebar: () => void;
    isSlim: () => boolean;
    isSlimPlus: () => boolean;
    isHorizontal: () => boolean;
    isDesktop: () => boolean;
    isSidebarActive: () => boolean;
    breadcrumbs?: Breadcrumb[];
    setBreadcrumbs: React.Dispatch<React.SetStateAction<Breadcrumb[]>>;
}